refactor(logcat): split query bar
